% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoAnnotate.R
\name{glycoAnnotationsCollapse}
\alias{glycoAnnotationsCollapse}
\title{Collapse annotated m/z values to have one row per peak/feature}
\usage{
glycoAnnotationsCollapse(annotated_data, collapse_columns, noncollapse_columns)
}
\description{
\code{glycoAnnotationsCollapse()} collapses the output of
\link[GlycoAnnotateR]{glycoAnnotate} in the case of multiple annotations
per peak or feature so that there is one row per peak/feature with
multiple annotations comma-separated.
}
\section{Slots}{

\describe{
\item{\code{annotated_data}}{Dataframe annotated by \link[GlycoAnnotateR]{glycoAnnotate}
that has NOT been collapsed and has multiple annotations per peak/feature.}

\item{\code{collapse_columns}}{Names of columns to be pasted together before collapsing.
Suggested is molecule name and ion.}

\item{\code{noncollapse_columns}}{Names of columns that uniquely identify peaks and
that should be retained after collapsing - these are generally the column
names of your input dataframe before annotation.}
}}

\examples{
#annotate dataframe
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos", naming = "IUPAC")
annotated_data <- glycoAnnotate(data = data, param = gpp, error = 1.5, units = 'ppm', collapse = F)

#collapse multiple annotations
annotated_data_collapsed <- glycoAnnotationsCollapse(annotated_data = annotated_data, collapse_columns = c('IUPAC name', 'ion'), noncollapse_columns = c('mz', 'rt', 'sampleA', 'sampleB'))

}
\seealso{
\link[GlycoAnnotateR]{glycoPredictParam}

\link[GlycoAnnotateR]{glycoPredict}

\link[GlycoAnnotateR]{glycoAnnotate}
}
