% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoPredict.R
\name{glycoPredict}
\alias{glycoPredict}
\title{Predict masses and m/z values of theoretical glycans}
\usage{
glycoPredict(param)
}
\arguments{
\item{param}{A \code{glycoPredictParam} object. See \link[GlycoAnnotateR]{glycoPredictParam-class}}
}
\description{
\code{glycoPredict()} predicts all possible glycan within the
constraints set by the \code{glycoPredictParam} object.
}
\details{
\code{glycoPredict()} is used to predict masses and mass to charge ratios of all theoretically
possible glycans within a set of constraining parameters (defined in the
\code{glycoPredictParam} object). This package was written
for annotation of mass spec data (especially LC-MS) but if used for
other purposes either ionisation mode and very wide scan ranges can be given.
The function works by sourcing a python file and then using the function
encoded in the python script.
}
\examples{
pgp <- glycoPredictParam()
pgp@dp <- c(1,7)
pgp@polarity <- 'neg'
pgp@scan_range <- c(150, 1300)
pgp@modifications <- c('sulphate', 'carboxylicacid')
pgp@double_sulphate <- TRUE
predicted.df <- glycoPredict(param = pgp)

}
\seealso{
glycoAnnotateR::glycoPredictParam()
}
