% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{glycoPredictParam-class}
\alias{glycoPredictParam-class}
\alias{glycoPredictParam}
\title{glycoPredictParam class}
\description{
An S4 class that contains the parameters for \link[GlycoAnnotateR]{glycoPredict()} function.
}
\details{
\code{glycoPredict()} is used to predict masses and mass to charge ratios of all theoretically
possible glycans within a set of constraining parameters (defined in the
\code{glycoPredictParam} object). This package was written
for annotation of mass spec data (especially LC-MS) but if used for
other purposes either ionisation mode and very wide scan ranges can be given.
The function works by sourcing a python file and then using the function
encoded in the python script.
}
\section{Slots}{

\describe{
\item{\code{dp}}{Degree of polymerisation range. It must be a numeric vector of length 2. Default: c(1,6).}

\item{\code{polarity}}{Ionisation mode used. Accepts 'pos' or 'neg' (default).}

\item{\code{scan_range}}{Scan range used during MS. It must be a numeric vector of length 2. Default: c(175, 1400).}

\item{\code{pent_option}}{Logical. Should pentose monomers be included? Default: \code{FALSE}.}

\item{\code{modifications}}{Modifications to be considered. Any combination of 'carboxylicacid', 'sialicacid',
'phosphate', 'deoxy', 'nacetyl', 'omethyl', 'anhydrobridge', 'oacetyl',
'unsaturated', 'alditol', 'amino', 'dehydrated', 'sulphate', 'aminopentyllinker' or 'all' or 'none' (default)}

\item{\code{nmod_max}}{Maximum number of modifications per monomer, calculated by the number of modifications over the number of monomers (default 1).
Does not take into account unsaturated, alditol or dehydrated.}

\item{\code{double_sulphate}}{Logical. Can monomers be double-sulphated. If \code{TRUE}, nmod_max needs to have a value of at least 2.}

\item{\code{label}}{Are sugars labelled by reductive amination? Current supported labels are: "none", "procainamide","2-aminobenzoic acid",
"2-aminobenzamide", "1-phenyl-3-methyl-5-pyrazolone".}

\item{\code{ion_type}}{Ionisation type. Currently accepted ESI and MALDI. Impacts ions.}

\item{\code{naming}}{Notation for molecule names. Uses commonly accepted abbreviations. Possibilities: 'IUPAC' (default), 'Oxford', 'GlycoCT'}

\item{\code{glycan_linkage}}{Option to implement filters for O- and N-glycans. Possibilities: 'none' (default), 'nglycan' or 'oglycan'.}
}}

\examples{
gpp <- glycoPredictParam()
gpp@dp <- c(1,7)
gpp@polarity <- 'neg'
gpp@scan_range <- c(150, 1300)
gpp@modifications <- c('sulphate', 'carboxylicacid')
gpp@double_sulphate <- TRUE
predicted.df <- glycoPredict(param = gpp)

}
\seealso{
GlycoAnnotateR::glycoPredict()
}
