% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGlycans.R
\name{predictGlycans}
\alias{predictGlycans}
\title{Function to predict glycan masses and m/z values}
\usage{
predictGlycans(
  dp1,
  dp2,
  ESI_mode,
  scan_range1,
  scan_range2,
  pent_option = NULL,
  modifications = NULL,
  nmod_max = NULL,
  double_sulphate = NULL,
  label = NULL
)
}
\arguments{
\item{dp1}{Minimum degree of polymerisation. Requires an integer.}

\item{dp2}{Minimum degree of polymerisation. Requires an integer.}

\item{ESI_mode}{ESI mode used. Accepts 'pos' or 'neg'}

\item{scan_range1}{Lower end of scan range. Requires an integer.}

\item{scan_range2}{Upper end of scan range. Requires an integer.}

\item{pent_option}{Should pentose monomers be included? 1 = yes, 0 = no. Defaults to no.}

\item{modifications}{Modifications to be considered. Any combination of 'carboxyl', 'phosphate', 'deoxy', 'nacetyl', 'omethyl', 'anhydrobridge', 'oacetyl', 'unsaturated', 'alditol', 'amino', 'dehydrated', 'sulphate' or 'all' or 'none' (default)}

\item{nmod_max}{Maximum number of modifications per monomer on average (default 1). Does not take into account unsaturated, alditol or dehydrated.}

\item{double_sulphate}{Can monomers be double-sulphated: 0 for no (default), 1 for yes. For yes you MUST give a value of at least 2 to nmod_max.}

\item{label}{Are sugars labelled? Currently only accepts none (default, leave out) or 'procainamide'.}
}
\description{
This function will predict all possible glycan molecules within the constraints of the parameters.
}
\examples{
mz_predicted <- predictGlycans(dp1 = 1, dp2 = 8, ESI_mode = 'neg', scan_range1 = 175, scan_range2 = 1400, pent_option = 1, modifications = c("sulphate", "deoxy", "carboxyl"),nmod_max = 2, double_sulphate = 1)
}
