% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoAnnotate.R
\name{glycoAnnotate}
\alias{glycoAnnotate}
\title{Annotate m/z values with glycan compositions based on theoretical values}
\usage{
glycoAnnotate(
  data,
  mz_column = "mz",
  mzmin_column = NULL,
  mzmax_column = NULL,
  pred_table = NULL,
  param = NULL,
  collapse = F,
  collapse_columns = NULL,
  error = 3,
  error_units = "ppm"
)
}
\description{
\code{glycoAnnotate()} annotates peaks or features in MS data,
using either a pre-generated table by \link[GlycoAnnotateR]{glycoPredict} or
by generating a new table.
}
\section{Slots}{

\describe{
\item{\code{data}}{Dataframe containing data to be annotated. For example,
feature dataframe from XCMS pre-processing (LC-MS or direct inject)
or features from Cardinal (MALDI).}

\item{\code{mz_column}}{Name of column containing m/z values.}

\item{\code{mzmin_column}}{OPTIONAL: Name of column containing minimum m/z data values.
If supplied, will do overlap-overlap matching. Generally only if mzmin and mzmax
values generated during peak picking. If not provided, mz value will be annotated
if within range of theoretical mz +- error.}

\item{\code{mzmax_column}}{OPTIONAL: Name of column containing maximum m/z data values.
If supplied, will do overlap-overlap matching. Generally only if mzmin and mzmax
values generated during peak picking.If not provided, mz value will be annotated
if within range of theoretical mz +- error.}

\item{\code{pred_table}}{Table generated previously by \link[GlycoAnnotateR]{glycoPredict}.}

\item{\code{param}}{\link[GlycoAnnotateR]{glycoPredictParam} object for generation of table
of theoretical mz values for annotation.}

\item{\code{collapse}}{Logical. If \code{TRUE}, annotations will be 'collapsed' so that multiple
annotations for one mz will be in the same row, comma separated (nrow of output is in
this case equal to nrow of input data). If \code{FALSE} (default), it is possible
that rows in the input dataframe are repeated with different annotations. The
information on annotations is more detailed in this case. Collapsing can also be done
afterwards on the output using \link[GlycoAnnotateR]{glycoAnnotationsCollapse}.}

\item{\code{collapse_columns}}{Columns to be pasted together before collapsing.
Only needed if \code{collapse=TRUE} and non-default columns wanted - default is
molecule name and ion. If prediction table provided to \code{pred_table} instead of
\code{param}, column names are required.}

\item{\code{error}}{Numeric value - error used to create window for matching. mz values
will be matched against theoretical mzs +- error.}

\item{\code{error_units}}{Units for error - can be 'ppm' or 'Da'}
}}

\examples{

#with prediction parameters
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos", naming = "IUPAC")
annotated_data <- glycoAnnotate(data = data, param = gpp, error = 1.5, units = 'ppm', collapse = T)

#with prediction table
gpp <- glycoPredictParam(dp = c(1, 8), modifications = "deoxy", polarity = "pos",  naming = "IUPAC")
pred_table <- glycoPredict(param = gpp)
annotated_data <- glycoAnnotate(data = data, pred_table = pred_table, error = 1.5, units = 'ppm', collapse = T, collapse_columns = c("IUPAC name", "ion"))

}
\seealso{
glycoAnnotateR::glycoPredict()

glycoAnnotateR::glycoPredictParam()
}
