% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoMS2Annotate.R
\name{glycoMS2Annotate}
\alias{glycoMS2Annotate}
\title{Annotate m/z values of fragements in MS2 spectra from glycans}
\usage{
glycoMS2Annotate(
  precursorAnnotations,
  precursorAnnotations_sep = NULL,
  ms2spectra,
  error = 3,
  error_units = "ppm",
  ion_type = "ESI",
  polarity = "neg",
  nmod_max = 1,
  double_sulfate = FALSE,
  label = "none"
)
}
\description{
\code{glycoMS2Annotate()} annotates fragments in MS2 spectra
from precursors annotated as glycans by \link[GlycoAnnotateR]{glycoAnnotate}.
}
\section{Slots}{

\describe{
\item{\code{precursorAnnotations}}{Vector of precursor annotations including IUPAC name,
ion and DP, separated by a colon (:) unless otherwise specified. Assumes
multiple annotations (if multiple annotations assigned to one precursor) are
separated by a comma!}

\item{\code{precursorAnnotations_sep}}{Non-default separator of IUPAC name, ion and DP
in \code{precursorAnnotations}}

\item{\code{ms2spectra}}{MS2 spectra from fragmentation of annotated precursors. MUST be
a dataframe with an 'mz' column for fragment ions and a 'precursorAnnotation'
column with precursor annotations (matching those in \code{precursorAnnotations})}

\item{\code{ion_type}}{Ionisation type. Currently accepted ESI and MALDI. Impacts ions.}

\item{\code{error}}{Numeric value - error used to create window for matching. mz values
will be matched against theoretical mzs +/- error.}

\item{\code{error_units}}{Units for error - can be 'ppm' or 'Da'}

\item{\code{nmod_max}}{Maximum number of modifications per monomer, calculated by the number of modifications over the number of monomers (default 1).
Does not take into account unsaturated, alditol or dehydrated.}

\item{\code{double_sulfate}}{Logical. Can monomers be double-sulfated. If \code{TRUE}, nmod_max needs to have a value of at least 2.}

\item{\code{label}}{Are sugars labelled by reductive amination? Current supported labels are: "none", "procainamide","2-aminobenzoic acid",
"2-aminobenzamide", "1-phenyl-3-methyl-5-pyrazolone".}
}}

\seealso{
\link[GlycoAnnotateR]{glycoPredictParam}

\link[GlycoAnnotateR]{glycoPredict}

\link[GlycoAnnotateR]{glycoAnnotate}

\link[GlycoAnnotateR]{glycoMS2Extract}
}
