% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoMS2Extract.R
\name{glycoMS2Extract}
\alias{glycoMS2Extract}
\title{MS/MS Spectra Extraction from annotated features}
\usage{
glycoMS2Extract(data_ms2, data_features, annotations)
}
\arguments{
\item{data_ms2}{(MSnbase::MSnExp), (MSnbase::OnDiskMSnExp) or (xcms::XCMSnExp) object with MS/MS spectra.}

\item{data_features}{(xcms::XCMSnExp) with MS1 features defined by XCMS processing
that were annotated by \link[GlycoAnnotateR]{glycoAnnotate}.}

\item{annotations}{Output of \link[GlycoAnnotateR]{glycoAnnotate}. It needs to be a
\code{data.frame} with numerical columns named "mz" and "rt" (mz columns contains \emph{m/z}
of the features, not of the annotations!).}
}
\value{
It returns a MSpectra object with all msLevel=2 spectra whose
precursors are the features annotated by \link[GlycoAnnotateR]{glycoAnnotate} function.
}
\description{
This function extracts MS/MS spectra associated with
features annotated by \link[GlycoAnnotateR]{glycoAnnotate} function.
}
\seealso{
glycoAnnotateR::glycoPredict()

glycoAnnotateR::glycoPredictParam()
}
